+++ addOns/fuzz/src/main/java/org/zaproxy/zap/extension/fuzz/ExtensionFuzz.java
@@ -128,15 +128,6 @@ public class ExtensionFuzz extends ExtensionAdaptor {
 
     private static final Logger LOGGER = LogManager.getLogger(ExtensionFuzz.class);
 
-    private static final ImageIcon SCRIPT_PAYLOAD_GENERATOR_ICON =
-            new ImageIcon(
-                    ExtensionFuzz.class.getResource(
-                            "resources/icons/script-payload-generator.png"));
-    private static final ImageIcon SCRIPT_PAYLOAD_PROCESSOR_ICON =
-            new ImageIcon(
-                    ExtensionFuzz.class.getResource(
-                            "resources/icons/script-payload-processor.png"));
-
     public static final String NAME = "ExtensionFuzz";
 
     private static final String JBROFUZZ_CATEGORY_PREFIX = "jbrofuzz";
@@ -308,7 +299,7 @@ public class ExtensionFuzz extends ExtensionAdaptor {
                     new ScriptType(
                             ScriptStringPayloadGenerator.TYPE_NAME,
                             "fuzz.payloads.script.type.payloadgenerator",
-                            SCRIPT_PAYLOAD_GENERATOR_ICON,
+                            createIcon("script-payload-generator.png"),
                             true,
                             true);
             extensionScript.registerScriptType(scriptTypeGenerator);
@@ -320,7 +311,7 @@ public class ExtensionFuzz extends ExtensionAdaptor {
                     new ScriptType(
                             ScriptStringPayloadProcessor.TYPE_NAME,
                             "fuzz.payloads.script.type.payloadprocessor",
-                            SCRIPT_PAYLOAD_PROCESSOR_ICON,
+                            createIcon("script-payload-processor.png"),
                             true,
                             true);
             extensionScript.registerScriptType(scriptTypeProcessor);
@@ -331,6 +322,13 @@ public class ExtensionFuzz extends ExtensionAdaptor {
 
         clientMessagePanelManager = new MessagePanelManager();
         serverMessagePanelManager = new MessagePanelManager();
+    }
+
+    private ImageIcon createIcon(String path) {
+        if (hasView()) {
+            return new ImageIcon(getClass().getResource("resources/icons/" + path));
+        }
+        return null;
     }
 
     @Override