+++ addOns/scripts/src/main/java/org/zaproxy/zap/extension/scripts/ScriptsTreeCellRenderer.java
@@ -124,7 +124,7 @@ public class ScriptsTreeCellRenderer extends DefaultTreeCellRenderer {
 
             if (node.isRoot() || node.getParent().isRoot()) {
                 // Top 2 levels use same icon .. for now ;)
-                setIcon(DisplayUtils.getScaledIcon(ExtensionScriptsUI.ICON));
+                setIcon(DisplayUtils.getScaledIcon(ExtensionScriptsUI.getIcon()));
 
             } else if (userObject != null && userObject instanceof ScriptWrapper) {
                 OverlayIcon icon;
@@ -152,13 +152,16 @@ public class ScriptsTreeCellRenderer extends DefaultTreeCellRenderer {
                         } else {
                             icon =
                                     new OverlayIcon(
-                                            DisplayUtils.getScaledIcon(ExtensionScriptsUI.ICON));
+                                            DisplayUtils.getScaledIcon(
+                                                    ExtensionScriptsUI.getIcon()));
                         }
                     } else if (engine.getIcon() != null) {
                         icon = new OverlayIcon(DisplayUtils.getScaledIcon(engine.getIcon()));
                     } else {
                         // Default to the blank script
-                        icon = new OverlayIcon(DisplayUtils.getScaledIcon(ExtensionScriptsUI.ICON));
+                        icon =
+                                new OverlayIcon(
+                                        DisplayUtils.getScaledIcon(ExtensionScriptsUI.getIcon()));
                     }
                     if (script.isChanged() && !node.isTemplate()) {
                         icon.add(DisplayUtils.getScaledIcon(PENCIL_OVERLAY_ICON));