+++ addOns/sequence/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Maintenance changes.
 - Update minimum ZAP version to 2.12.0.
 
+### Fixed
+- Prevent exception if no display (Issue 3978).
+
 ## [6] - 2021-10-07
 ### Added
 - Add info and repo URLs.